Use the publicly-available cortex-m-rt crate
