+++ .cargo/config
@@ -3,5 +3,5 @@ target = "thumbv7em-none-eabihf"
 
 [target.thumbv7em-none-eabihf]
 rustflags = [
-    "-C", "link-arg=-Tlink.x",
+    "-C", "link-arg=-Tt4link.x",
 ]
