--- cortex-m-rt-patch/Cargo.toml
@@ -1,10 +1,0 @@
-# NOTE: This is a temporary patch for `cortex-m-rt` - see the README.
-
-[package]
-name = "cortex-m-rt"
-version = "0.6.999+teensy4"
-authors = ["mitchmindtree <mitchell.nordine@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-teensy4-rt = { path = "../teensy4-rt" }
