--- cortex-m-rt-patch/src/lib.rs
@@ -1,3 +1,0 @@
-#![no_std]
-
-pub use teensy4_rt::*;
