+++ examples/bsp/Cargo.toml
@@ -11,6 +11,7 @@ path = "../../"
 
 [dependencies]
 cortex-m = "0.6.2"
+cortex-m-rt = "0.6.13"
 embedded-hal = "0.2.4"
 log = "0.4.8"
 nb = "0.1.2"
