+++ examples/bsp/src/bin/gpt.rs
@@ -10,7 +10,7 @@ extern crate panic_halt;
 
 use bsp::hal::gpt;
 use bsp::interrupt;
-use bsp::rt::{entry, interrupt};
+use cortex_m_rt::{entry, interrupt};
 use teensy4_bsp as bsp;
 
 use core::time::Duration;
