+++ examples/bsp/src/bin/led.rs
@@ -7,8 +7,8 @@
 
 extern crate panic_halt;
 
-use bsp::rt::entry;
 use cortex_m::asm::wfi;
+use cortex_m_rt::entry;
 use teensy4_bsp as bsp;
 
 use embedded_hal::digital::v2::OutputPin;
