+++ examples/bsp/src/bin/spi.rs
@@ -23,7 +23,7 @@
 
 extern crate panic_halt;
 
-use bsp::rt::entry;
+use cortex_m_rt::entry;
 use teensy4_bsp as bsp;
 
 use embedded_hal::{blocking::spi::Transfer, digital::v2::OutputPin};
