+++ examples/bsp/src/bin/usb.rs
@@ -10,7 +10,7 @@
 
 extern crate panic_halt;
 
-use bsp::rt;
+use cortex_m_rt as rt;
 use teensy4_bsp as bsp;
 
 #[rt::entry]
