+++ src/rt/cache.rs
