--- teensy4-rt/src/fault.rs
@@ -1,60 +1,0 @@
-use core::fmt;
-
-/// Registers stacked (pushed into the stack) during an exception
-#[derive(Clone, Copy)]
-#[repr(C)]
-pub struct ExceptionFrame {
-    /// (General purpose) Register 0
-    pub r0: u32,
-
-    /// (General purpose) Register 1
-    pub r1: u32,
-
-    /// (General purpose) Register 2
-    pub r2: u32,
-
-    /// (General purpose) Register 3
-    pub r3: u32,
-
-    /// (General purpose) Register 12
-    pub r12: u32,
-
-    /// Linker Register
-    pub lr: u32,
-
-    /// Program Counter
-    pub pc: u32,
-
-    /// Program Status Register
-    pub xpsr: u32,
-}
-
-impl fmt::Debug for ExceptionFrame {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        struct Hex(u32);
-        impl fmt::Debug for Hex {
-            fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-                write!(f, "0x{:08x}", self.0)
-            }
-        }
-        f.debug_struct("ExceptionFrame")
-            .field("r0", &Hex(self.r0))
-            .field("r1", &Hex(self.r1))
-            .field("r2", &Hex(self.r2))
-            .field("r3", &Hex(self.r3))
-            .field("r12", &Hex(self.r12))
-            .field("lr", &Hex(self.lr))
-            .field("pc", &Hex(self.pc))
-            .field("xpsr", &Hex(self.xpsr))
-            .finish()
-    }
-}
-
-#[doc(hidden)]
-#[cfg_attr(target_arch = "arm", link_section = ".HardFault.default")]
-#[no_mangle]
-pub unsafe extern "C" fn HardFault_(_: &ExceptionFrame) -> ! {
-    loop {
-        core::sync::atomic::spin_loop_hint();
-    }
-}
